IRMO.jl

Documentation for IRMO.jl

IRMO.H — Method

CDF of the binomial distribution

IRMO._sample — Method

Functions for sampling from the simplex

IRMO.chop_pt — Method

Remove numerical zeros

IRMO.delete_zeros — Method

Remove rows of zeros

IRMO.get_cooperation — Method

Calculate the rate of cooperation

IRMO.get_cooperation_n — Method

Calculate the rate of cooperation for norm competition

IRMO.get_cooperation_nf — Method

Calculate the rate of cooperation for strategy competition under fixed mixture of norms

IRMO.get_cooperation_normInd — Method

Calculate the rate of cooperation for strategy and norm independent evolution.

IRMO.get_cooperation_s — Method

Calculate the rate of cooperation

IRMO.get_full_dynamics — Method

Reputation and replicator dynamics

IRMO.get_full_dynamics_n — Method

Reputation and replicator dynamics for norm competition

IRMO.get_full_dynamics_nf — Method

Reputation and replicator dynamics for strategy competition under fixed mixture of norms

IRMO.get_full_dynamics_s — Method

Reputation and replicator dynamics

IRMO.get_norm — Method

Get the social norm matrix

IRMO.get_rep — Method

Get reputation dynamics

IRMO.get_rep_n — Method

Get reputation dynamics for norm competition

IRMO.get_rep_nf — Method

Get reputation dynamics for strategy competition under fixed mixture of norms

IRMO.get_rep_s — Method

Get replicator dynamics for time trajectories

IRMO.get_time_dynamics — Method

Get replicator dynamics for time trajectories

IRMO.get_time_dynamics_n — Method

Get replicator dynamics for time trajectories for norm competition

IRMO.get_time_dynamics_nf — Method

Get replicator dynamics for time trajectories for strategy competition under fixed mixture of norms

IRMO.get_time_dynamics_s — Method

Get reputation dynamics

IRMO.refine_normInd — Method

Expand integration to a lower tolerance for strategy and norm independent evolution.

IRMO.refine_steady_states — Method

Expand integration to a lower tolerance.

IRMO.refine_steady_states_n — Method

Expand integration to a lower tolerance for norm competition.

IRMO.refine_steady_states_nf — Method

Expand integration to a lower tolerance for strategy competition under fixed mixture of norms.

IRMO.refine_steady_states_s — Method

Expand integration to a lower tolerance.

IRMO.steady_states_detIC — Method

Integrate numerically until a steady state is reached. The initial frequencies are deterministic, evenly spaced, and determined by the global variable grid.

IRMO.steady_states_detIC_n — Method

Integrate numerically until a steady state is reached. The initial frequencies are deterministic, evenly spaced, and determined by the global variable grid. Function fitted for norm competition.

IRMO.steady_states_detIC_nf — Method

Integrate numerically until a steady state is reached. The initial frequencies are deterministic, evenly spaced, and determined by the global variable grid. Function fitted for strategy competition under fixed mixture of norms.

IRMO.steady_states_detIC_s — Method

Integrate numerically until a steady state is reached. The initial frequencies are deterministic, evenly spaced, and determined by the global variable grid.

IRMO.steady_states_normInd — Method

Integrate numerically until a steady state is reached. The initial frequencies are deterministic, evenly spaced, and determined by the global variable grid. Function fitted for strategy and norm independent evolution.

IRMO.steady_states_rndIC — Method

Integrate numerically until a steady state is reached. The initial frequencies are sampled uniformly at random.

IRMO.steady_states_rndIC_n — Method

Integrate numerically until a steady state is reached. The initial frequencies are sampled uniformly at random. Function fitted for norm competition.

IRMO.steady_states_rndIC_s — Method

Integrate numerically until a steady state is reached. The initial frequencies are sampled uniformly at random.

IRMO.trajectories_detIC — Method

Integrate numerically until a steady state is reached, saving the full time trajectory.